Move common functions only used by browser_action to browser_action
