+++ src/common/functions.ts
@@ -1,12 +1,8 @@
-/* global window,Node */
-
-import BookmarkTreeNode = chrome.bookmarks.BookmarkTreeNode;
-import {BookmarkOpener, BookmarkOpeningDisposition} from "./BookmarkOpener";
-
 export function nothing(e: Event) {
   e.preventDefault();
   e.stopPropagation();
   e.stopImmediatePropagation();
+
   return false;
 }
 
@@ -37,45 +33,4 @@ export function translateDocument(document: Document): void {
       translatableElement.innerHTML = translation;
     }
   });
-}
-
-export function getElementData(element: HTMLElement, key: string): string {
-  return element.getAttribute(`data-${key}`);
-}
-
-export function setElementData(element: HTMLElement, key: string, value: string): void {
-  element.setAttribute(`data-${key}`, value);
-}
-
-export function getAncestorsWithClass(element: Element, className: string): HTMLElement[] {
-  const parents = [];
-
-  if (!(element.parentNode instanceof Element)) {
-    return parents;
-  }
-
-  if (element.parentNode.classList.contains(className)) {
-    parents.push(element.parentNode);
-  }
-
-  return parents.concat(getAncestorsWithClass(element.parentNode, className));
-}
-
-export function elementIndex(element: Element): number {
-  if (!element || typeof element.parentNode === 'undefined') {
-    return null;
-  }
-  const parent = element.parentNode;
-  const children = parent.childNodes;
-  let i = 0;
-  for (let j = 0; j < children.length; j++) {
-    if (children[j].nodeType === Node.TEXT_NODE) {
-      continue;
-    }
-    if (element === children[j]) {
-      return i;
-    }
-    i++;
-  }
-  return -1;
 }
